% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{cotramICP}
\alias{cotramICP}
\title{ICP for count transformation models}
\usage{
cotramICP(formula, data, env, ...)
}
\arguments{
\item{formula}{Formula including response and shift terms}

\item{data}{Data.frame containing response and explanatory variables}

\item{env}{Formula specifying the exogenous environment variables}

\item{...}{
  Arguments passed on to \code{\link[=dicp]{dicp}}
  \describe{
    \item{\code{modFUN}}{Model function from package 'tram', i.e. \code{BoxCox},
\code{Colr}, \code{Polr}, \code{Lm}, \code{Coxph}, \code{Survreg},
\code{Lehmann}. Standard implementations \code{lm}, \code{glm}, and
\code{\link[MASS]{polr}} are also supported.}
    \item{\code{verbose}}{Logical, whether output should be verbose (default \code{TRUE})}
    \item{\code{type}}{Character, type of invariance (\code{"residual"}, \code{"wald"},
or \code{"partial"})}
    \item{\code{test}}{Character, type of test to be used (\code{"HSIC"}, \code{"t.test"},
\code{"var.test"}, \code{"wald"}) or custom function for testing
invariance.}
    \item{\code{controls}}{Controls for the used tests and the overall procedure,
see \code{dicp_controls}}
    \item{\code{alpha}}{Level of invariance test, default}
  }}
}
\value{
Object of type \code{"dICP"}. See \code{\link[tramicp]{dicp}}
}
\description{
ICP for count transformation models
}
\examples{
set.seed(13312)
d <- dgp_dicp(mod = "cotram")
cotramICP(Y ~ X2, data = d, env = ~ E, type = "wald", test = "wald")
cotramICP(Y ~ X2, data = d, env = ~ E, type = "residual", test = "HSIC")
cotramICP(Y ~ X2, data = d, env = ~ E, type = "residual", test = "indep")
cotramICP(Y ~ X2, data = d, env = ~ E, type = "confint")
glmICP(Y ~ X2, data = d, env = ~ E, family = "poisson", test = "indep")

}
