% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{dicp_controls}
\alias{dicp_controls}
\title{TRAMICP Controls}
\usage{
dicp_controls(
  type = "residual",
  test = "independence",
  baseline_fixed = TRUE,
  alpha = 0.05,
  method = "gamma",
  kernel = c("gaussian", "discrete"),
  B = 499,
  vcov = "vcov",
  teststat = "maximum",
  distribution = "asymptotic",
  xtrafo = trafo,
  ytrafo = trafo,
  residuals = "residuals",
  stop_if_empty_set_invariant = getOption("stop_if_empty_set_invariant", default =
    FALSE)
)
}
\arguments{
\item{type}{Character, type of invariance (\code{"residual"}, \code{"wald"},
or \code{"partial"}).}

\item{test}{Character, type of test to be used (\code{"HSIC"}, \code{"t.test"},
\code{"var.test"}, \code{"wald"}) or custom function for testing
invariance.}

\item{baseline_fixed}{logical; whether or not the baseline transformation
is allowed to vary with the environments. Only takes effect when
\code{type} is one of \code{"wald"}, \code{"confint"}, or \code{"mcheck"}.}

\item{alpha}{Level of invariance test, default \code{0.05}.}

\item{method}{See \code{\link[dHSIC]{dhsic.test}}.}

\item{kernel}{See \code{\link[dHSIC]{dhsic.test}}.}

\item{B}{See \code{\link[dHSIC]{dhsic.test}}.}

\item{vcov}{(Name of) function for computing the variance-covariance matrix of a model.}

\item{teststat}{See \code{\link[coin]{independence_test}}.}

\item{distribution}{See \code{\link[coin]{independence_test}}.}

\item{xtrafo}{See \code{\link[coin]{independence_test}}.}

\item{ytrafo}{See \code{\link[coin]{independence_test}}.}

\item{residuals}{(Name of) function for computing model residuals.}
}
\value{
List of dicp controls
}
\description{
TRAMICP Controls
}
