% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tram-sim.R
\name{.tram_setup}
\alias{.tram_setup}
\title{Setting up a general transformation model}
\usage{
.tram_setup(
  type = c("ordered", "continuous", "survival", "count"),
  distr = c("Logistic", "MinExtrVal", "MaxExtrVal", "Normal"),
  basis = c("bernstein", "linear", "log"),
  K = 6,
  supp = .supp(type),
  bounds = .bounds(type),
  extrapolate = FALSE,
  log_first = FALSE,
  numeric_vars = c("X1", "X2"),
  cfb = c(-2, -1),
  cfx = c(0.6, -0.3),
  factor_vars = "E",
  factor_levels = c(1, 2),
  interacting = FALSE,
  ia_fac = 0.7,
  add = c(0, 0)
)
}
\arguments{
\item{type}{Response type}

\item{distr}{Inverse link}

\item{basis}{Basis type}

\item{K}{Number of outcome classes or basis functions}

\item{supp}{Support of response}

\item{bounds}{Bounds of response}

\item{extrapolate}{See \code{\link[basefun]{Bernstein_basis}}}

\item{log_first}{See \code{\link[basefun]{Bernstein_basis}}}

\item{numeric_vars}{Names of numeric variables}

\item{cfb}{Baseline coef}

\item{cfx}{Shift coefs}

\item{factor_vars}{Names of factor variables}

\item{factor_levels}{Levels of factors}

\item{interacting}{Toggle varying baseline}

\item{ia_fac}{Scale coef}

\item{add}{See \code{\link[tram]{tram}}}
}
\value{
Object of class \code{"mlt"}
}
\description{
Setting up a general transformation model
}
