% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramicp.R
\name{invariant_sets}
\alias{invariant_sets}
\title{Return invariant sets}
\usage{
invariant_sets(object, with_pvalues = FALSE)
}
\arguments{
\item{object}{Object of class \code{"dICP"}.}

\item{with_pvalues}{Logical; whether to also return p-values of invariance
tests for the non-rejected sets.}
}
\value{
Returns vector of all non-rejected sets. With
   \code{with_pvalues = TRUE}, a named vector of p-values is returned.
   Returns \code{named numeric(0)} if there are no invariant sets.
}
\description{
Return invariant sets
}
