% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{LmICP}
\alias{LmICP}
\title{ICP for normal linear regression}
\usage{
LmICP(formula, data, env, ...)
}
\arguments{
\item{formula}{Formula including response and shift terms}

\item{data}{Data.frame containing response and explanatory variables}

\item{env}{Character, name of environmental variable}

\item{...}{
  Arguments passed on to \code{\link[=dicp]{dicp}}
  \describe{
    \item{\code{modFUN}}{Model function from package 'tram', i.e. \code{BoxCox},
\code{Colr}, \code{Polr}, \code{Lm}, \code{Coxph}, \code{Survreg},
\code{Lehmann}. Standard implementation \code{lm} is also supported.}
    \item{\code{verbose}}{Logical, whether output should be verbose (default \code{TRUE})}
    \item{\code{type}}{Character, type of invariance (\code{"residual"}, \code{"wald"},
or \code{"partial"})}
    \item{\code{test}}{Character, type of test to be used (\code{"HSIC"}, \code{"t.test"},
\code{"var.test"}, \code{"wald"}) or custom function for testing
invariance.}
    \item{\code{trt}}{Character, supply only when \code{type = "partial"}, treatment
variable. Ignored otherwise}
    \item{\code{baseline_fixed}}{Logical, whether baseline transformation is fixed
(\code{TRUE}) or allowed to vary across environments (\code{FALSE}).
Defaults to \code{TRUE}, i.e. a fixed baseline transformation}
    \item{\code{alpha}}{Level of invariance test, default}
    \item{\code{controls}}{Controls for the used tests, see \code{dicp_controls}}
  }}
}
\value{
Object of type \code{"dICP"}. See \code{\link[tramicp]{dicp}}
}
\description{
ICP for normal linear regression
}
\examples{
d <- dgp_dicp(mod = "lm")
LmICP(Y ~ X2, data = d, env = "E", type = "wald", test = "wald")
LmICP(Y ~ X2, data = d, env = "E", type = "residual", test = "HSIC")
LmICP(Y ~ X2, data = d, env = "E", type = "residual", test = "indep")
LmICP(Y ~ X2, data = d, env = "E", type = "confint")
dicp(Y ~ X2, data = d, env = "E", type = "confint", modFUN = lm)

}
