% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{PolrICP}
\alias{PolrICP}
\title{ICP for cumulative ordinal regression}
\usage{
PolrICP(
  formula,
  data,
  env,
  verbose = TRUE,
  type = "residual",
  test = "independence",
  controls = NULL,
  alpha = 0.05,
  baseline_fixed = TRUE,
  greedy = FALSE,
  ...
)
}
\arguments{
\item{formula}{Formula including response and shift terms}

\item{data}{Data.frame containing response and explanatory variables}

\item{env}{Formula specifying the exogenous environment variables}

\item{verbose}{Logical, whether output should be verbose (default \code{TRUE})}

\item{type}{Character, type of invariance (\code{"residual"}, \code{"wald"},
or \code{"partial"})}

\item{test}{Character, type of test to be used (\code{"HSIC"}, \code{"t.test"},
\code{"var.test"}, \code{"wald"}) or custom function for testing
invariance.}

\item{controls}{Controls for the used tests and the overall procedure,
see \code{dicp_controls}}

\item{alpha}{Level of invariance test, default}

\item{baseline_fixed}{Fixed baseline transformation, see
\code{\link[tramicp]{dicp_controls}}.}

\item{greedy}{Logical, whether to perform a greedy version of ICP (default is
\code{FALSE})}

\item{...}{
  Arguments passed on to \code{\link[=dicp]{dicp}}
  \describe{
    \item{\code{modFUN}}{Model function from package 'tram', i.e. \code{BoxCox},
\code{Colr}, \code{Polr}, \code{Lm}, \code{Coxph}, \code{Survreg},
\code{Lehmann}. Standard implementations \code{lm}, \code{glm}, and
\code{\link[MASS]{polr}} are also supported.}
  }}
}
\value{
Object of type \code{"dICP"}. See \code{\link[tramicp]{dicp}}
}
\description{
ICP for cumulative ordinal regression
}
\examples{
d <- dgp_dicp(mod = "polr")
PolrICP(Y ~ X2, data = d, env = ~ E, type = "wald", test = "wald")
PolrICP(Y ~ X2, data = d, env = ~ E, type = "residual", test = "HSIC")
PolrICP(Y ~ X2, data = d, env = ~ E, type = "residual", test = "indep")
PolrICP(Y ~ X2, data = d, env = ~ E, type = "confint")

}
