% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dICP}
\alias{print.dICP}
\title{Distributional ICP}
\usage{
\method{print}{dICP}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"dICP"}}

\item{...}{Currently ignored}
}
\value{
Invisible \code{NULL}
}
\description{
Distributional ICP
}
