% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dICPtest}
\alias{print.dICPtest}
\title{Distributional ICP test}
\usage{
\method{print}{dICPtest}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"dICPtest"}}

\item{...}{Currently ignored}
}
\value{
Numeric vector of p-values from invariance tests
}
\description{
Distributional ICP test
}
