% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dICP}
\alias{summary.dICP}
\title{Summary method dICP}
\usage{
\method{summary}{dICP}(object, print_all = FALSE, digits = 3, ...)
}
\arguments{
\item{object}{Object of class \code{"dICP"}}

\item{print_all}{Print p-values for all tested sets}

\item{digits}{Numeric, number of digits to print, default is 3}

\item{...}{Currently ignored}
}
\value{
Invisible \code{NULL}
}
\description{
Summary method dICP
}
