% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{SurvregICP}
\alias{SurvregICP}
\title{ICP for parametric survival models}
\usage{
SurvregICP(
  formula,
  data,
  env,
  verbose = TRUE,
  type = "residual",
  test = "gcm.test",
  controls = NULL,
  alpha = 0.05,
  baseline_fixed = TRUE,
  greedy = FALSE,
  max_size = NULL,
  mandatory = NULL,
  ...
)
}
\arguments{
\item{formula}{A \code{formula} including response and covariate terms.}

\item{data}{A \code{data.frame} containing response and explanatory variables.}

\item{env}{A \code{formula} specifying the environment variables (see details).}

\item{verbose}{Logical, whether output should be verbose (default \code{TRUE}).}

\item{type}{Character, type of invariance (\code{"residual"}, \code{"wald"},
or \code{"partial"}).}

\item{test}{Character, type of test to be used (\code{"HSIC"}, \code{"t.test"},
\code{"var.test"}, \code{"wald"}) or custom function for testing
invariance.}

\item{controls}{Controls for the used tests and the overall procedure,
see \code{dicp_controls}.}

\item{alpha}{Level of invariance test, default \code{0.05}.}

\item{baseline_fixed}{Fixed baseline transformation, see
\code{\link[tramicp]{dicp_controls}}.}

\item{greedy}{Logical, whether to perform a greedy version of ICP (default is
\code{FALSE}).}

\item{max_size}{Numeric; maximum support size.}

\item{mandatory}{A \code{formula} containing mandatory covariates, i.e.,
covariates which by a priori knowledge are believed to be parents
of the response or are in another way required for the model to be
valid (for instance, random effects in a mixed model).}

\item{...}{
  Arguments passed on to \code{\link[=dicp]{dicp}}
  \describe{
    \item{\code{modFUN}}{Model function from package 'tram', i.e.,
\code{\link[tram]{BoxCox}}, \code{\link[tram]{Colr}},
\code{\link[tram]{Polr}}, \code{\link[tram]{Lm}},
\code{\link[tram]{Coxph}}, \code{\link[tram]{Survreg}},
\code{\link[tram]{Lehmann}}. Standard implementations
\code{\link[stats]{lm}}, \code{\link[stats]{glm}},
\code{\link[survival]{survreg}}, \code{\link[survival]{coxph}},
and \code{\link[MASS]{polr}} are also supported. See the corresponding
alias \code{<model_name>ICP}, e.g., \code{\link{PolrICP}}.}
  }}
}
\value{
Object of type \code{"dICP"}. See \code{\link[tramicp]{dicp}}
}
\description{
ICP for parametric survival models
}
\examples{
set.seed(123)
d <- dgp_dicp(mod = "weibull")
SurvregICP(Y ~ X1 + X2 + X3, data = d, env = ~ E)

}
