% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramicp.R
\name{dicp_controls}
\alias{dicp_controls}
\title{Controls}
\usage{
dicp_controls(
  type = "residual",
  test = "independence",
  baseline_fixed = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{type}{Character, type of invariance (\code{"residual"}, \code{"wald"},
or \code{"partial"})}

\item{test}{Character, type of test to be used (\code{"HSIC"}, \code{"t.test"},
\code{"var.test"}, \code{"wald"}) or custom function for testing
invariance.}

\item{baseline_fixed}{logical; whether or not the baseline transformation
is allowed to vary with the environments. Only takes effect when
\code{type} is one of \code{"wald"}, \code{"confint"}, or \code{"mcheck"}.}

\item{alpha}{Level of invariance test, default}
}
\value{
List of dicp controls
}
\description{
Controls
}
