% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tram-sim.R
\name{.tram_setup}
\alias{.tram_setup}
\title{Setting up a general transformation model}
\usage{
.tram_setup(
  type = c("ordered", "continuous", "survival", "count"),
  distr = c("Logistic", "MinExtrVal", "MaxExtrVal", "Normal"),
  basis = c("bernstein", "linear", "log"),
  K = 6,
  supp = .supp(type),
  bounds = .bounds(type),
  extrapolate = FALSE,
  log_first = FALSE,
  numeric_vars = c("X1", "X2"),
  cfb = c(-2, -1),
  cfx = c(0.6, -0.3),
  factor_vars = "E",
  factor_levels = c(1, 2),
  interacting = FALSE,
  ia_fac = 0.7,
  add = c(0, 0)
)
}
\arguments{
\item{type}{Response type}

\item{distr}{Inverse link}

\item{basis}{Basis type}

\item{K}{Number of outcome classes or basis functions}

\item{supp}{Support of response}

\item{bounds}{Bounds of response}

\item{extrapolate}{See \code{\link[basefun]{Bernstein_basis}}}

\item{log_first}{See \code{\link[basefun]{Bernstein_basis}}}

\item{numeric_vars}{Names of numeric variables}

\item{cfb}{Baseline coef}

\item{cfx}{Shift coefs}

\item{factor_vars}{Names of factor variables}

\item{factor_levels}{Levels of factors}

\item{interacting}{Toggle varying baseline}

\item{ia_fac}{Scale coef}

\item{add}{See \code{\link[tram]{tram}}}
}
\value{
Object of class \code{"mlt"}
}
\description{
Setting up a general transformation model
}
\examples{
# Polr model
m <- .tram_setup()
plot(m, newdata = data.frame(X1 = 0, X2 = 0), type = "trafo", col = 1)

# Weibull model
m <- .tram_setup(type = "survival", cfb = c(-1,1), basis = "log")
plot(m, newdata = data.frame(X1 = 0, X2 = 0), type = "trafo", col = 1)

# Cox model
m <- .tram_setup(type = "survival", basis = "bernstein")
plot(m, newdata = data.frame(X1 = -3, X2 = -3), type = "distribution", col = 1)

# Cotram
m <- .tram_setup(type = "count", basis = "bernstein")
plot(m, newdata = data.frame(X1 = -1, X2 = -1), type = "distribution", col = 1)

# BoxCox
m <- .tram_setup(type = "continuous", basis = "bernstein", distr = "Normal")
plot(m, newdata = data.frame(X1 = -1, X2 = -1), type = "distribution", col = 1)

# Small check
m <- .tram_setup(cfx = c(0, 0.5), type = "count", basis = "bernstein")
nd <- data.frame(X1 = runif(tn <- 1e4), X2 = runif(tn),
                 E = factor(1 + rbinom(tn, 1, 0.5)))
nd$Y <- as.integer(trtf:::.R2vec(simulate(m, newdata = nd)))
confint(cotram::cotram(Y ~ X1 + X2, data = nd, prob = 0.999))

}
